+++ CHANGELOG.md
@@ -36,3 +36,4 @@ We follow [Semantic Versions](https://semver.org/).
 ## Version 0.4.3
 
 - Fix resource placement bug
+- Change typings for `init_resources` and `shutdown_resources`. Now they always return `Awaitable`.